FossilOrigin-Name: 9e6cd5ec27709005cbb8e79d647bf166a10842a80469e3fda912db8faa1405cc

+++ src/ams/delivery/trymail/trymail.c
@@ -2946,7 +2946,7 @@ char *DefaultDomain;
 	AuthFrom = atoi(AuthString);
 	if (AuthFrom == 0) {
 	    KS = NULL;
-	    wperr = cwp_Search(wpCD, STok, -1, MatchIDNamePart, &outQuality, &KS);
+	    wperr = cwp_Search(wpCD, STok, (long)-1, MatchIDNamePart, &outQuality, &KS);
 	    if (wperr == wperr_NoError && KS != NULL) {
 		NPAmax = NPAix = -100000;
 		for (Ix = 0; Ix < KS->KeyCount; ++Ix) {
